Update once.js
Updated onceClass to allow for a mixed (Array and Map) syntax as per the ngClass docs:

```once-class="[style4, {orange: warning}]"```

i'm aware its not perfect and needs refactoring, however it is a starting point to add the behaviour. 

refer to https://docs.angularjs.org/api/ng/directive/ngClass for the angular doc on ngClass.